refactor: Remove InstanceInfoRepository, and cleanup
Code that needs instance info (e.g., to make UI decisions based on server
limits) should be going through the instance info attached to the
account, not querying `InstanceInfoRepository` directly.

Clean up the users.

Since `AccountManager` populates the local database with the instance info
and is the only writer there's no need for `InstanceInfoRepository` after
the users are cleaned up.
