+++ app/src/test/java/app/pachli/components/compose/ComposeActivityTest.kt
@@ -23,7 +23,6 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import app.pachli.PachliApplication
 import app.pachli.R
 import app.pachli.core.data.repository.AccountManager
-import app.pachli.core.data.repository.InstanceInfoRepository
 import app.pachli.core.database.AppDatabase
 import app.pachli.core.model.Draft
 import app.pachli.core.model.InstanceInfo.Companion.DEFAULT_CHARACTERS_RESERVED_PER_URL
@@ -111,9 +110,6 @@ class ComposeActivityTest {
 
     @Inject
     lateinit var accountManager: AccountManager
-
-    @Inject
-    lateinit var instanceInfoRepository: InstanceInfoRepository
 
     @Inject
     lateinit var db: AppDatabase
@@ -320,7 +316,7 @@ class ComposeActivityTest {
     fun whenMaximumTootCharsIsPopulated_customLimitIsUsed() = runTest {
         val customMaximum = 1000
         getInstanceCallback = { getInstanceWithCustomConfiguration(customMaximum, getCustomInstanceConfiguration(maximumStatusCharacters = customMaximum)) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
@@ -335,7 +331,7 @@ class ComposeActivityTest {
     fun whenOnlyLegacyMaximumTootCharsIsPopulated_customLimitIsUsed() = runTest {
         val customMaximum = 1000
         getInstanceCallback = { getInstanceWithCustomConfiguration(customMaximum) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
@@ -350,7 +346,7 @@ class ComposeActivityTest {
     fun whenOnlyConfigurationMaximumTootCharsIsPopulated_customLimitIsUsed() = runTest {
         val customMaximum = 1000
         getInstanceCallback = { getInstanceWithCustomConfiguration(null, getCustomInstanceConfiguration(maximumStatusCharacters = customMaximum)) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
@@ -365,7 +361,7 @@ class ComposeActivityTest {
     fun whenDifferentCharLimitsArePopulated_statusConfigurationLimitIsUsed() = runTest {
         val customMaximum = 1000
         getInstanceCallback = { getInstanceWithCustomConfiguration(customMaximum, getCustomInstanceConfiguration(maximumStatusCharacters = customMaximum * 2)) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
@@ -499,7 +495,7 @@ class ComposeActivityTest {
         val additionalContent = "Check out this @image #search result: "
         val customUrlLength = 16
         getInstanceCallback = { getInstanceWithCustomConfiguration(configuration = getCustomInstanceConfiguration(charactersReservedPerUrl = customUrlLength)) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
@@ -521,7 +517,7 @@ class ComposeActivityTest {
         val additionalContent = " Check out this @image #search result: "
         val customUrlLength = 18 // The intention is that this is longer than shortUrl.length
         getInstanceCallback = { getInstanceWithCustomConfiguration(configuration = getCustomInstanceConfiguration(charactersReservedPerUrl = customUrlLength)) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
@@ -542,7 +538,7 @@ class ComposeActivityTest {
         val additionalContent = " Check out this @image #search result: "
         val customUrlLength = 16
         getInstanceCallback = { getInstanceWithCustomConfiguration(configuration = getCustomInstanceConfiguration(charactersReservedPerUrl = customUrlLength)) }
-        instanceInfoRepository.reload(accountManager.activeAccount)
+        accountManager.refresh(pachliAccountId)
 
         rule.launch()
         dispatcher.scheduler.advanceUntilIdle()
